refactor: moved `CoinSelector`, `FundingError` and `UTXO` classes to a new file
