+++ lib/primitives/mtx.js
@@ -9,7 +9,6 @@
 
 const assert = require('bsert');
 const {encoding} = require('bufio');
-const {BufferMap} = require('buffer-map');
 const Script = require('../script/script');
 const TX = require('./tx');
 const Input = require('./input');
@@ -17,13 +16,12 @@ const Output = require('./output');
 const Coin = require('./coin');
 const Outpoint = require('./outpoint');
 const CoinView = require('../coins/coinview');
-const Address = require('./address');
 const consensus = require('../protocol/consensus');
 const policy = require('../protocol/policy');
-const Amount = require('../btc/amount');
 const Stack = require('../script/stack');
 const util = require('../utils/util');
 const {inspectSymbol} = require('../utils');
+const {CoinSelector, FundingError} = require('../wallet/coinselector');
 
 /**
  * MTX
@@ -1485,891 +1483,6 @@ class MTX extends TX {
   }
 }
 
-/**
- * UTXO (Coin with additional properties)
- */
-
-class UTXO {
-  /**
-   * Create a credit.
-   * @constructor
-   */
-
-  constructor() {
-    this.spendingSize = 0;
-    this.effectiveValue = 0;
-    this.index = -1;
-  }
-}
-
-/**
- * Coin Selector
- * @alias module:primitives.CoinSelector
- */
-
-class CoinSelector {
-  /**
-   * Create a coin selector.
-   * @constructor
-   * @param {TX} tx
-   * @param {Object?} options
-   */
-
-  constructor(tx, options) {
-    this.tx = tx.clone();
-    this.coins = [];
-    this.utxos = [];
-    this.outputValue = 0;
-    this.index = 0;
-    this.chosen = [];
-    this.change = 0;
-    this.fee = CoinSelector.MIN_FEE;
-
-    this.selection = 'value';
-    this.subtractFee = false;
-    this.subtractIndex = -1;
-    this.height = -1;
-    this.depth = -1;
-    this.hardFee = -1;
-    this.rate = CoinSelector.FEE_RATE;
-    this.maxFee = -1;
-    this.round = false;
-    this.changeAddress = null;
-    this.inputs = new BufferMap();
-    this.useSelectEstimate = false;
-
-    // Needed for size estimation.
-    this.getAccount = null;
-
-    this.injectInputs();
-
-    if (options)
-      this.fromOptions(options);
-  }
-
-  /**
-   * Initialize selector options.
-   * @param {Object} options
-   * @private
-   */
-
-  fromOptions(options) {
-    if (options.selection) {
-      assert(typeof options.selection === 'string');
-      this.selection = options.selection;
-    }
-
-    if (options.subtractFee != null) {
-      if (typeof options.subtractFee === 'number') {
-        assert(Number.isSafeInteger(options.subtractFee));
-        assert(options.subtractFee >= -1);
-        this.subtractIndex = options.subtractFee;
-        this.subtractFee = this.subtractIndex !== -1;
-      } else {
-        assert(typeof options.subtractFee === 'boolean');
-        this.subtractFee = options.subtractFee;
-      }
-    }
-
-    if (options.subtractIndex != null) {
-      assert(Number.isSafeInteger(options.subtractIndex));
-      assert(options.subtractIndex >= -1);
-      this.subtractIndex = options.subtractIndex;
-      this.subtractFee = this.subtractIndex !== -1;
-    }
-
-    if (options.height != null) {
-      assert(Number.isSafeInteger(options.height));
-      assert(options.height >= -1);
-      this.height = options.height;
-    }
-
-    if (options.confirmations != null) {
-      assert(Number.isSafeInteger(options.confirmations));
-      assert(options.confirmations >= -1);
-      this.depth = options.confirmations;
-    }
-
-    if (options.depth != null) {
-      assert(Number.isSafeInteger(options.depth));
-      assert(options.depth >= -1);
-      this.depth = options.depth;
-    }
-
-    if (options.hardFee != null) {
-      assert(Number.isSafeInteger(options.hardFee));
-      assert(options.hardFee >= -1);
-      this.hardFee = options.hardFee;
-    }
-
-    if (options.rate != null) {
-      assert(Number.isSafeInteger(options.rate));
-      assert(options.rate >= 0);
-      this.rate = options.rate;
-    }
-
-    if (options.maxFee != null) {
-      assert(Number.isSafeInteger(options.maxFee));
-      assert(options.maxFee >= -1);
-      this.maxFee = options.maxFee;
-    }
-
-    if (options.round != null) {
-      assert(typeof options.round === 'boolean');
-      this.round = options.round;
-    }
-
-    if (options.useSelectEstimate != null) {
-      assert(typeof options.useSelectEstimate === 'boolean');
-      this.useSelectEstimate = options.useSelectEstimate;
-    }
-
-    if (options.changeAddress) {
-      const addr = options.changeAddress;
-      if (typeof addr === 'string') {
-        this.changeAddress = Address.fromString(addr);
-      } else {
-        assert(addr instanceof Address);
-        this.changeAddress = addr;
-      }
-    }
-
-    if (options.getAccount) {
-      assert(typeof options.getAccount === 'function');
-      this.getAccount = options.getAccount;
-    }
-
-    if (options.inputs) {
-      assert(Array.isArray(options.inputs));
-      for (let i = 0; i < options.inputs.length; i++) {
-        const prevout = options.inputs[i];
-        assert(prevout && typeof prevout === 'object');
-        const {hash, index} = prevout;
-        assert(Buffer.isBuffer(hash));
-        assert(typeof index === 'number');
-        this.inputs.set(Outpoint.toKey(hash, index), i);
-      }
-    }
-
-    return this;
-  }
-
-  /**
-   * Attempt to inject existing inputs.
-   * @private
-   */
-
-  injectInputs() {
-    if (this.tx.inputs.length > 0) {
-      for (let i = 0; i < this.tx.inputs.length; i++) {
-        const {prevout} = this.tx.inputs[i];
-        this.inputs.set(prevout.toKey(), i);
-      }
-    }
-  }
-
-  /**
-   * Initialize the selector with coins to select from.
-   * @param {Coin[]} coins
-   */
-
-  async init(coins) {
-    this.coins = coins.slice();
-    for (let i = 0, n = coins.length; i < n; i++) {
-      const coin = coins[i];
-      if (this.isSpendable(coin)) {
-        const utxo = new UTXO();
-        utxo.index = i;
-        utxo.spendingSize = await coin.estimateSpendingSize(this.getAccount);
-        utxo.effectiveValue = coin.value - this.getFee(utxo.spendingSize);
-        this.utxos.push(utxo);
-      }
-    }
-    this.utxos.sort((a, b) => b.effectiveValue - a.effectiveValue);
-    this.outputValue = this.tx.getOutputValue();
-    this.index = 0;
-    this.chosen = [];
-    this.change = 0;
-    this.fee = CoinSelector.MIN_FEE;
-    this.tx.inputs.length = 0;
-
-    switch (this.selection) {
-      case 'all':
-      case 'random':
-        this.coins.sort(sortRandom);
-        break;
-      case 'age':
-        this.coins.sort(sortAge);
-        break;
-      case 'value':
-        this.coins.sort(sortValue);
-        break;
-      default:
-        throw new FundingError(`Bad selection type: ${this.selection}.`);
-    }
-  }
-
-  /**
-   * Calculate total value required.
-   * @returns {Amount}
-   */
-
-  total() {
-    if (this.subtractFee)
-      return this.outputValue;
-    return this.outputValue + this.fee;
-  }
-
-  /**
-   * Test whether the selector has
-   * completely funded the transaction.
-   * @returns {Boolean}
-   */
-
-  isFull() {
-    return this.tx.getInputValue() >= this.total();
-  }
-
-  /**
-   * Test whether a coin is spendable
-   * with regards to the options.
-   * @param {Coin} coin
-   * @returns {Boolean}
-   */
-
-  isSpendable(coin) {
-    if (this.tx.view.hasEntry(coin))
-      return false;
-
-    if (this.height === -1)
-      return true;
-
-    if (coin.coinbase) {
-      if (coin.height === -1)
-        return false;
-
-      if (this.height + 1 < coin.height + consensus.COINBASE_MATURITY)
-        return false;
-
-      return true;
-    }
-
-    if (this.depth === -1)
-      return true;
-
-    const depth = coin.getDepth(this.height);
-
-    if (depth < this.depth)
-      return false;
-
-    return true;
-  }
-
-  /**
-   * Get the current fee based on a size.
-   * @param {Number} size
-   * @returns {Amount}
-   */
-
-  getFee(size) {
-    // This is mostly here for testing.
-    // i.e. A fee rounded to the nearest
-    // kb is easier to predict ahead of time.
-    if (this.round) {
-      const fee = policy.getRoundFee(size, this.rate);
-      return Math.min(fee, CoinSelector.MAX_FEE);
-    }
-
-    const fee = policy.getMinFee(size, this.rate);
-    return Math.min(fee, CoinSelector.MAX_FEE);
-  }
-
-  /**
-   * Fund the transaction with more
-   * coins if the `output value + fee`
-   * total was updated.
-   */
-
-  fund() {
-    // Ensure all preferred inputs first.
-    if (this.inputs.size > 0) {
-      const coins = [];
-
-      for (let i = 0; i < this.inputs.size; i++)
-        coins.push(null);
-
-      for (const coin of this.coins) {
-        const {hash, index} = coin;
-        const key = Outpoint.toKey(hash, index);
-        const i = this.inputs.get(key);
-
-        if (i != null) {
-          coins[i] = coin;
-          this.inputs.delete(key);
-        }
-      }
-
-      if (this.inputs.size > 0)
-        throw new Error('Could not resolve preferred inputs.');
-
-      for (const coin of coins) {
-        this.tx.addCoin(coin);
-        this.chosen.push(coin);
-      }
-    }
-
-    while (this.index < this.coins.length) {
-      const coin = this.coins[this.index++];
-
-      if (!this.isSpendable(coin))
-        continue;
-
-      this.tx.addCoin(coin);
-      this.chosen.push(coin);
-
-      if (this.selection === 'all')
-        continue;
-
-      if (this.isFull())
-        break;
-    }
-  }
-
-  /**
-   * Initiate selection from `coins`.
-   * @param {Coin[]} coins
-   * @returns {CoinSelector}
-   */
-
-  async select(coins) {
-    await this.init(coins);
-
-    if (this.hardFee !== -1) {
-      this.selectHard();
-    } else if (this.useSelectEstimate) {
-      // This is potentially asynchronous:
-      // it may invoke the size estimator
-      // required for redeem scripts (we
-      // may be calling out to a wallet
-      // or something similar).
-      await this.selectEstimate();
-    } else {
-      await this.attemptSelection();
-    }
-
-    if (!this.isFull()) {
-      // Still failing to get enough funds.
-      throw new FundingError(
-        'Not enough funds.',
-        this.tx.getInputValue(),
-        this.total());
-    }
-
-    // How much money is left after filling outputs.
-    this.change = this.tx.getInputValue() - this.total();
-
-    return this;
-  }
-
-  /**
-   * Initialize selection based on size estimate.
-   */
-
-  async selectEstimate() {
-    // Set minimum fee and do
-    // an initial round of funding.
-    this.fee = CoinSelector.MIN_FEE;
-    this.fund();
-
-    // Add dummy output for change.
-    const change = new Output();
-
-    if (this.changeAddress) {
-      change.script.fromAddress(this.changeAddress);
-    } else {
-      // In case we don't have a change address,
-      // we use a fake p2pkh output to gauge size.
-      change.script.fromPubkeyhash(Buffer.allocUnsafe(20));
-    }
-
-    this.tx.outputs.push(change);
-
-    // Keep recalculating the fee and funding
-    // until we reach some sort of equilibrium.
-    do {
-      const size = await this.tx.estimateSize(this.getAccount);
-
-      this.fee = this.getFee(size);
-
-      if (this.maxFee > 0 && this.fee > this.maxFee)
-        throw new FundingError('Fee is too high.');
-
-      // Failed to get enough funds, add more coins.
-      if (!this.isFull())
-        this.fund();
-    } while (!this.isFull() && this.index < this.coins.length);
-  }
-
-  /**
-   * Initiate selection based on a hard fee.
-   */
-
-  selectHard() {
-    this.fee = Math.min(this.hardFee, CoinSelector.MAX_FEE);
-    this.fund();
-  }
-
-  /**
-   * Initiate multi algorithm coin selection.
-   */
-
-  async attemptSelection() {
-    // calculate cost of creating and spending a change output
-    // add dummy output for change.
-    const change = new Output();
-
-    if (this.changeAddress) {
-      change.script.fromAddress(this.changeAddress);
-    } else {
-      // In case we don't have a change address,
-      // we use a fake p2pkh output to gauge size.
-      change.script.fromPubkeyhash(Buffer.allocUnsafe(20));
-    }
-
-    const changeOutputSize = change.getSize();
-    const changeOutputCoin = new Coin();
-    changeOutputCoin.script = change.script;
-    const changeSpendingSize =
-      await changeOutputCoin.estimateSpendingSize(this.getAccount);
-    const costOfSpendingChange =
-      changeSpendingSize * CoinSelector.LONG_TERM_FEERATE;
-
-    // cost of change = fee paid to create change +
-    // fee paid to spend that change in future
-    const costOfChange = this.getFee(changeOutputSize) + costOfSpendingChange;
-
-    let target = this.outputValue;
-    let size = await this.tx.estimateSize();
-    const fee = this.getFee(size);
-    target += fee;
-
-    // try to find a changeless solution using BnB
-    let result = await this.selectBnB(target, costOfChange);
-    if (result && result.length > 0) {
-      // we have found a changeless solution
-      for (const i of result) {
-        const coin = this.coins[i];
-        this.tx.addCoin(coin);
-        this.chosen.push(coin);
-      }
-      // update fee
-      size = await this.tx.estimateSize();
-      this.fee = this.getFee(size);
-      return;
-    }
-
-    // fallback to other algorithms as we can't find a changeless solution
-    // update target to include cost of producing change
-    target += this.getFee(changeOutputSize);
-
-    let bestSelection = [];
-    let bestWaste = CoinSelector.MAX_MONEY;
-    // ideally change value should be atleast 100 times of
-    // cost of spending it in future. We will try to find such a solution.
-    const usefulChange = 100 * costOfSpendingChange;
-    // find a solution of target using Lowest Larger
-    result = this.selectLowestLarger(target);
-    if (result.length > 0) {
-      const currWaste = this.getWaste(result, costOfChange, target);
-      if (currWaste < bestWaste) {
-        bestSelection = result;
-        bestWaste = currWaste;
-      }
-    }
-    // try to find a solution with useful change
-    result = this.selectLowestLarger(target + usefulChange);
-    if (result.length > 0) {
-      const currWaste = this.getWaste(result, costOfChange,
-        target + usefulChange);
-      if (currWaste < bestWaste) {
-        bestSelection = result;
-        bestWaste = currWaste;
-      }
-    }
-
-    if (bestSelection && bestSelection.length > 0) {
-      for (const i of bestSelection) {
-        const coin = this.coins[i];
-        this.tx.addCoin(coin);
-        this.chosen.push(coin);
-      }
-      // add change output
-      this.tx.outputs.push(change);
-      // update fee
-      size = await this.tx.estimateSize();
-      this.fee = this.getFee(size);
-    }
-  }
-
-  /**
-   * Initiate selection using Single Random Draw selection.
-   * @param {Number} target - Selection target
-   * @returns {Number[]} selected - array of indicies of selected coins
-   */
-
-  selectSRD(target) {
-    const coins = this.coins.map((coin, index) => {
-      coin.pos = index;
-      return coin;
-    });
-    coins.sort(() => Math.random() - 0.5);
-
-    let index = 0;
-    const selected = [];
-    let selectedValue = 0;
-
-    while (index < coins.length) {
-      const coin = coins[index];
-      selectedValue += coin.effectiveValue;
-      selected.push(coin.pos);
-      if (selectedValue >= target)
-        return selected;
-      index++;
-    }
-
-    return [];
-  }
-
-  /**
-   * Initiate selection using Branch and Bound selection.
-   * @param {Number} target - Selection target
-   * @param {Number} costOfChange - Cost of producing and spending change
-   * @returns {Number[]} selected - array of indicies of selected coins
-   */
-
-  async selectBnB(target, costOfChange) {
-    const selected = [];
-    let currValue = 0;
-    let totalValue = 0;
-    let index = 0;
-    for (const utxo of this.utxos)
-      totalValue += utxo.effectiveValue;
-
-    // we don't have enough funds
-    if (totalValue < target) {
-      throw new FundingError(
-        'Not enough funds.',
-        totalValue,
-        target);
-    }
-
-    let tries = 100000;
-    let bestWaste = CoinSelector.MAX_MONEY;
-    let bestSelection = [];
-    // perform depth-first search for choosing UTXOs
-    while (tries-- > 0) {
-      let backtrack = false;
-      // conditions for backtracking
-      // 1. cannot reach target with remaining amount
-      // 2. selected value is greater than upper bound
-      if (currValue + totalValue < target
-        || currValue > target + costOfChange) {
-        backtrack = true;
-      } else if (currValue >= target) {
-        // we have a solution, we will compare it with current best solution
-        // we will try to find the solution with minimum waste
-        const currWaste = this.getWaste(selected, 0, target);
-        if (currWaste <= bestWaste) {
-          bestSelection = selected.slice();
-          bestWaste = currWaste;
-          if (bestWaste <= 0)
-            break;
-        }
-        backtrack = true;
-      }
-
-      // backtracking here
-      if (backtrack) {
-        // we have backtracked to the first UTXO,
-        // all branches are traversed, we are done here
-        if (selected.length === 0)
-          break;
-
-        // Add omitted UTXOs back before traversing
-        // the omission branch of last included UTXO
-        // we are using a while loop here beacuse
-        // there can be more than one UTXOs that were omitted
-        while (--index > selected[selected.length - 1]) {
-          totalValue += this.utxos[index].effectiveValue;
-        }
-
-        // Remove last included UTXO from selected list.
-        currValue -= this.utxos[index].effectiveValue;
-        selected.pop();
-      } else {
-        const utxo = this.utxos[index];
-        // remove this UTXO from total available amount
-        totalValue -= utxo.effectiveValue;
-        // if this UTXO is the first one or
-        // if the previous index is included or
-        // if this UTXO's value is different from the previous one
-        if  (selected.length === 0
-          || index - 1 === selected[selected.length - 1]
-          || utxo.effectiveValue !== this.utxos[index - 1].effectiveValue) {
-          selected.push(index);
-          currValue += this.utxos[index].effectiveValue;
-        }
-      }
-      index++;
-    }
-
-    return bestSelection;
-  }
-
-  /**
-   * Initiate selection using Lowest Larger selection algorithm.
-   * @param {Number} target - Selection target
-   * @returns {Number[]} selected - array of indicies of selected coins
-   */
-
-  selectLowestLarger(target) {
-    // while target is greater than
-    // the largest coin we have, we
-    // will keep selecting the largest coin
-
-    this.index = 0;
-    const selected = [];
-    let effectiveValue = this.utxos[this.index].effectiveValue;
-    while (target >= effectiveValue) {
-      // update target, select current coin and increment index
-      selected.push(this.index);
-      target -= effectiveValue;
-      this.index++;
-
-      if (this.index === this.utxos.length)
-        break;
-
-      effectiveValue = this.utxos[this.index].effectiveValue;
-    }
-
-    if (target > 0 && this.index !== this.utxos.length) {
-      // now we are sure that target < largest unselected coin
-      // we will perform Binary search to find the smallest coin
-      // which is greater than target value
-
-      const closestCoinIndex = this.findLowestLarger(target);
-      target -= this.utxos[closestCoinIndex].effectiveValue;
-      selected.push(closestCoinIndex);
-    }
-
-    return target > 0 ? [] : selected;
-  }
-
-  /**
-   * Find smallest coin greater than
-   * the target using binary search
-   * @param {Number} target
-   * @returns {Number} index
-   */
-
-  findLowestLarger(target) {
-    let i = this.index;
-    let j = this.utxos.length - 1;
-    // Corner cases
-    // if target < smallest coin available
-    // return the smallest coin
-    if (target <= this.utxos[j].effectiveValue)
-      return j;
-
-    // begin binary search
-    let mid = 0;
-    while (i < j) {
-      mid = Math.floor((i + j) / 2);
-
-      // calculate effective value of coin at mid
-      const effectiveValue = this.utxos[mid].effectiveValue;
-
-      if (effectiveValue === target)
-        return mid;
-
-      // if target is less than coin at mid
-      // then search in left part of array
-      if (target < effectiveValue) {
-        if (mid < this.utxos.length - 1) {
-          // calculate effective value of coin at mid + 1
-          const nextEffectiveValue = this.utxos[mid + 1].effectiveValue;
-          // if target is greater than next
-          // effective value, then return mid
-          // coin at mid > target > coin at mid + 1
-          if (target > nextEffectiveValue)
-            return mid;
-        }
-        // repeat for right half
-        i = mid + 1;
-      } else {
-        // repeat for left half
-        j = mid;
-      }
-    }
-
-    // only single element left after search
-    return mid;
-  }
-
-  /**
-   * Calculate effectiveValue of a coin
-   * effectiveValue = coin.value - cost of spending that coin
-   * @param {Coin} coin - A coin object
-   * @returns {Number} - effective value of the coin
-   */
-
-  async getEffectiveValue(coin) {
-    const spendingSize = await coin.estimateSpendingSize(this.getAccount);
-    const effectiveValue = coin.value - this.getFee(spendingSize);
-    return effectiveValue;
-  }
-
-  /**
-   * Calculate waste for a selection
-   * @param {Number[]} selected - indicies of selected coins
-   * @param {Number} costOfChange - the cost of making change and spending it
-   * @param {Number} target - selection target
-   * @returns {Number} - waste
-   */
-
-  getWaste(selected, costOfChange, target) {
-    let waste = 0;
-    let selectedAmount = 0;
-    // calculate current fee rate in sats/byte
-    const currentFeeRate = this.getFee(1);
-
-    for (const i of selected) {
-      const utxo = this.utxos[i];
-      selectedAmount += utxo.effectiveValue;
-      // Consider the cost of spending an input now vs in the future
-      waste += (currentFeeRate - CoinSelector.LONG_TERM_FEERATE)
-        * utxo.spendingSize;
-    }
-
-    if (costOfChange > 0) {
-      // Consider the cost of making change and spending it in the future
-      waste += costOfChange;
-    } else {
-      // if costOfChange is not set,
-      // consider the excess we are throwing away to fees
-      waste += selectedAmount - target;
-    }
-
-    return waste;
-  }
-}
-
-/**
- * Default fee rate
- * for coin selection.
- * @const {Amount}
- * @default
- */
-
-CoinSelector.FEE_RATE = 10000;
-
-/**
- * Minimum fee to start with
- * during coin selection.
- * @const {Amount}
- * @default
- */
-
-CoinSelector.MIN_FEE = 10000;
-
-/**
- * Maximum fee to allow
- * after coin selection.
- * @const {Amount}
- * @default
- */
-
-CoinSelector.MAX_FEE = consensus.COIN / 10;
-
-/**
- * Maximum possible supply of Bitcoin
- * @const {Amount}
- * @default
- */
-
-CoinSelector.MAX_MONEY = 21000000 * 100000000;
-
-/**
- * Long term feerate (sats/vbyte)
- * @const {Amount}
- * @default
- */
-
-CoinSelector.LONG_TERM_FEERATE = 5;
-
-/**
- * Funding Error
- * An error thrown from the coin selector.
- * @ignore
- * @extends Error
- * @property {String} message - Error message.
- * @property {Amount} availableFunds
- * @property {Amount} requiredFunds
- */
-
-class FundingError extends Error {
-  /**
-   * Create a funding error.
-   * @constructor
-   * @param {String} msg
-   * @param {Amount} available
-   * @param {Amount} required
-   */
-
-  constructor(msg, available, required) {
-    super();
-
-    this.type = 'FundingError';
-    this.message = msg;
-    this.availableFunds = -1;
-    this.requiredFunds = -1;
-
-    if (available != null) {
-      this.message += ` (available=${Amount.btc(available)},`;
-      this.message += ` required=${Amount.btc(required)})`;
-      this.availableFunds = available;
-      this.requiredFunds = required;
-    }
-
-    if (Error.captureStackTrace)
-      Error.captureStackTrace(this, FundingError);
-  }
-}
-
-/*
- * Helpers
- */
-
-function sortAge(a, b) {
-  a = a.height === -1 ? 0x7fffffff : a.height;
-  b = b.height === -1 ? 0x7fffffff : b.height;
-  return a - b;
-}
-
-function sortRandom(a, b) {
-  return Math.random() > 0.5 ? 1 : -1;
-}
-
-function sortValue(a, b) {
-  if (a.height === -1 && b.height !== -1)
-    return 1;
-
-  if (a.height !== -1 && b.height === -1)
-    return -1;
-
-  return b.value - a.value;
-}
-
 function sortInputs(a, b) {
   return a.compare(b);
 }
@@ -2384,7 +1497,6 @@ function sortOutputs(a, b) {
 
 exports = MTX;
 exports.MTX = MTX;
-exports.Selector = CoinSelector;
 exports.FundingError = FundingError;
 
 module.exports = exports;
