+++ lib/wallet/wallet.js
@@ -31,6 +31,7 @@ const consensus = require('../protocol/consensus');
 const {Mnemonic} = HD;
 const {inspectSymbol} = require('../utils');
 const {BufferSet} = require('buffer-map');
+const {CoinSelector} = require('./coinselector');
 
 /**
  * Wallet
@@ -1142,7 +1143,7 @@ class Wallet extends EventEmitter {
       getAccount: this.getAccountByAddress.bind(this)
     });
 
-    assert(mtx.getFee() <= MTX.Selector.MAX_FEE, 'TX exceeds MAX_FEE.');
+    assert(mtx.getFee() <= CoinSelector.MAX_FEE, 'TX exceeds MAX_FEE.');
   }
 
   /**
@@ -1317,8 +1318,8 @@ class Wallet extends EventEmitter {
 
     let fee = tx.getMinFee(null, rate);
 
-    if (fee > MTX.Selector.MAX_FEE)
-      fee = MTX.Selector.MAX_FEE;
+    if (fee > CoinSelector.MAX_FEE)
+      fee = CoinSelector.MAX_FEE;
 
     if (oldFee >= fee)
       throw new Error('Fee is not increasing.');
