+++ test/util/memwallet.js
@@ -15,6 +15,7 @@ const {BloomFilter} = require('bfilter');
 const KeyRing = require('../../lib/primitives/keyring');
 const Outpoint = require('../../lib/primitives/outpoint');
 const Coin = require('../../lib/primitives/coin');
+const {CoinSelector} = require('../../lib/wallet/coinselector');
 
 class MemWallet {
   constructor(options) {
@@ -401,7 +402,7 @@ class MemWallet {
 
     await this.fund(mtx, options);
 
-    assert(mtx.getFee() <= MTX.Selector.MAX_FEE, 'TX exceeds MAX_FEE.');
+    assert(mtx.getFee() <= CoinSelector.MAX_FEE, 'TX exceeds MAX_FEE.');
 
     mtx.sortMembers();
